diagnostic/requirements.txt
paho-mqtt
requests
numpy
